--- packages/react/src/components/Callout/README.stories.mdx
@@ -1,67 +1,0 @@
-import { Props, Description } from '@storybook/addon-docs/blocks';
-import contributing from '../../../../../docs/contributing-license.md';
-import Callout from './Callout';
-
-# Callout
-
-## Getting started
-
-Here's a quick example to get you started.
-
-##### CSS
-
-```css
-// yourapplication.scss
-@import '@carbon/type/scss/font-face/mono';
-@import '@carbon/type/scss/font-face/sans';
-@include carbon--font-face-mono();
-@include carbon--font-face-sans();
-@import '@carbon/ibmdotcom-styles/scss/components/callout/callout';
-```
-
-> 💡 Only import fonts once per usage. Don't forget to import the Callout styles
-> from
-> [@carbon/ibmdotcom-styles](https://github.com/carbon-design-system/ibm-dotcom-library/blob/master/packages/styles).
-
-##### JS
-
-```javascript
-import React from 'react';
-import ReactDOM from 'react-dom';
-import { Callout } from '@carbon/ibmdotcom-react';
-import 'yourapplication.scss';
-
-function App() {
-  return (
-    <Callout>
-      <p>hello world</p>
-    </Callout>
-  );
-}
-
-ReactDOM.render(<App />, document.querySelector('#app'));
-```
-
-Add the following line on your `env` file at the root of your project,
-[See more details](https://github.com/carbon-design-system/ibm-dotcom-library/tree/master/packages/styles#usage).
-
-```
-  SASS_PATH=node_modules:src
-```
-
-> 💡 Don't forget to import the callout styles from
-> [@carbon/ibmdotcom-styles](https://github.com/carbon-design-system/ibm-dotcom-library/blob/master/packages/styles).
-
-## Props
-
-<Props of={Callout} />
-
-## Stable selectors
-
-| Name                      | Description             |
-| ------------------------- | ----------------------- |
-| `dds--callout__container` | Callout element wrapper |
-| `dds--column__column`     | Callout column wrapper  |
-| `dds--column__content`    | Callout content wrapper |
-
-<Description markdown={contributing} />
