--- packages/react/src/components/Callout/index.js
@@ -1,8 +1,0 @@
-/**
- * Copyright IBM Corp. 2016, 2020
- *
- * This source code is licensed under the Apache-2.0 license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-export { default as Callout } from './Callout';
