+++ packages/react/src/patterns/blocks/CalloutQuote/CalloutQuote.js
@@ -5,9 +5,10 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import { Callout, Quote } from '../../../components';
+import Callout from '../../../internal/components/Callout/Callout';
 import { settings as ddsSettings } from '@carbon/ibmdotcom-utilities';
 import PropTypes from 'prop-types';
+import { Quote } from '../../../components/Quote';
 import React from 'react';
 import settings from 'carbon-components/es/globals/js/settings';
 const { stablePrefix } = ddsSettings;
