+++ packages/styles/scss/patterns/blocks/callout-quote/index.scss
@@ -6,5 +6,5 @@
  */
 
 @import '../../../globals/imports';
-@import '../../../components/callout/callout';
+@import '../../../internal/callout/callout';
 @import '../../../components/quote/quote';
