+++ packages/styles/scss/patterns/blocks/callout-with-media/index.scss
@@ -6,6 +6,6 @@
  */
 
 @import '../../../globals/imports';
-@import '../../../components/callout/callout';
+@import '../../../internal/callout/callout';
 @import '../content-block-simple/content-block-simple';
 @import './callout-with-media';
